requirements.txt
alpaca-py>=0.21.0
anthropic>=0.39.0
pandas>=2.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
yfinance>=0.2.0
